removed long on 10

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,8 +10,8 @@
 
 int _atoi(char *s)
 {
-	unsigned long int i;
-	unsigned long int l = strlen(s);
+	int i;
+	int l = strlen(s);
 
 	for (i = 0; i < l; i++)
 	{
